PIG-1135: skewed join partitioner returns negative partition index  (yinghe via olgan)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/pig/trunk@889305 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/mapReduceLayer/PigMapReduce.java
@@ -157,7 +157,7 @@ public class PigMapReduce {
 			Byte tupleValIdx = 3;
 
             Byte index = (Byte)tuple.get(0);
-			Byte partitionIndex = -1;
+			Integer partitionIndex = -1;
         	// for partitioning table, the partition index isn't present
 			if (tuple.size() == 3) {
 				//super.collect(oc, tuple);
@@ -165,7 +165,7 @@ public class PigMapReduce {
 				tupleKeyIdx--;
 				tupleValIdx--;
 			} else {
-				partitionIndex = (Byte)tuple.get(1);
+				partitionIndex = (Integer)tuple.get(1);
 			}
 
             PigNullableWritable key =
